# perf(ext/url): optimize UrlParts op serialization