@AaronO 2021-08-19
> Have you considered making UrlParts into `strurct UrlParts(Url)`, and then implementing `serde::Serialize` on it manually? All the serialization logic could be abstracted into the serializer.
> 
> LGTM though.

Yes, I patched the `url` crate so that it exposes its internal offsets. That's a more efficient serialization that I've prototyped locally, but it either requires patching `url` or doing pointer arithmetic to recompute those internal offsets.